+++ FirebaseRemoteConfig/Tests/Unit/RCNRemoteConfigTest.m
@@ -1137,7 +1137,7 @@ static NSString *UTCToLocal(NSString *utcTime) {
 }
 
 // Manage different bundle locations for Swift Package Manager, CocoaPods static, CocoaPods dynamic.
-- (void)setDefaultsFor:(FIRRemoteConfig *)config namespace:(NSString *)namespace {
+- (void)setDefaultsFor:(FIRRemoteConfig *)config {
 #if SWIFT_PACKAGE
   NSBundle *bundle = Firebase_RemoteConfigUnit_SWIFTPM_MODULE_BUNDLE();
   NSString *plistFile = [bundle pathForResource:@"Defaults-testInfo" ofType:@"plist"];
@@ -1145,11 +1145,7 @@ static NSString *UTCToLocal(NSString *utcTime) {
   NSBundle *bundle = [NSBundle mainBundle];
   NSString *plistFile = [bundle pathForResource:@"Defaults-testInfo" ofType:@"plist"];
   if (plistFile != nil) {
-    if (namespace) {
-      [config setDefaultsFromPlistFileName:@"Defaults-testInfo" namespace:namespace];
-    } else {
-      [config setDefaultsFromPlistFileName:@"Defaults-testInfo"];
-    }
+    [config setDefaultsFromPlistFileName:@"Defaults-testInfo"];
     return;
   }
   // We've linked dynamically and the plist file is in the test's bundle.
@@ -1161,17 +1157,13 @@ static NSString *UTCToLocal(NSString *utcTime) {
   }
 #endif
   NSDictionary *defaults = [[NSDictionary alloc] initWithContentsOfFile:plistFile];
-  if (namespace) {
-    [config setDefaults:defaults namespace:namespace];
-  } else {
-    [config setDefaults:defaults];
-  }
+  [config setDefaults:defaults];
 }
 
 - (void)testSetDefaultsFromPlist {
   for (int i = 0; i < RCNTestRCNumTotalInstances; i++) {
     FIRRemoteConfig *config = _configInstances[i];
-    [self setDefaultsFor:config namespace:nil];
+    [self setDefaultsFor:config];
     XCTAssertEqualObjects(_configInstances[i][@"lastCheckTime"].stringValue,
                           UTCToLocal(@"2016-02-28 18:33:31"));
     XCTAssertEqual(_configInstances[i][@"isPaidUser"].boolValue, YES);